feat: move arrays into dedicated data folder
