+++ src/Profile/Content/ProfileInfo/SharedFollowers.js
@@ -3,6 +3,7 @@ import { Link } from "react-router-dom";
 import styled from "styled-components";
 import person from "./person.svg";
 import Title from "./Title";
+import followers from "../../../data/followers";
 
 const Container = styled.div`
   display: flex;
@@ -34,41 +35,6 @@ const Avatar = styled.img`
 `;
 
 const StLink = styled(Link)``;
-
-const publicURL = process.env.PUBLIC_URL;
-
-const followers = [
-  {
-    id: 1,
-    name: "/mirhasanjamil75",
-    src: `${publicURL}/img/avatar-person-1.png`
-  },
-  {
-    id: 2,
-    name: "/alexdown",
-    src: `${publicURL}/img/avatar-person-2.png`
-  },
-  {
-    id: 3,
-    name: "/TerryCoopey",
-    src: `${publicURL}/img/avatar-person-3.png`
-  },
-  {
-    id: 4,
-    name: "/NiamhKeaneB_ID",
-    src: `${publicURL}/img/avatar-person-4.png`
-  },
-  {
-    id: 5,
-    name: "/venglarcik",
-    src: `${publicURL}/img/avatar-person-5.png`
-  },
-  {
-    id: 6,
-    name: "/katzedsgn",
-    src: `${publicURL}/img/avatar-person-6.png`
-  }
-];
 
 export default () => (
   <Container>
