+++ src/Profile/Content/ProfilePosts/Trend.js
@@ -2,6 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import { FormattedNumber } from "react-intl";
 import styled from "styled-components";
+import trends from "../../../data/trends";
 
 const Headline = styled.span`
   color: #1da1f2;
@@ -40,45 +41,6 @@ const Text = styled.p`
 `;
 
 const Counter = Text.extend``;
-
-export const trends = [
-  {
-    id: 1,
-    tag: "#BringYourDogToWorkDay",
-    text: null,
-    tweets: null
-  },
-  {
-    id: 2,
-    tag: "#FridayFeeling",
-    text: null,
-    tweets: 12100
-  },
-  {
-    id: 3,
-    tag: "#BrexitAnniversary",
-    text: "It’s one year since the UK voted to leave the European Union",
-    tweets: null
-  },
-  {
-    id: 4,
-    tag: "HMS Queen Elizabeth",
-    text: null,
-    tweets: 1036
-  },
-  {
-    id: 5,
-    tag: "Joe Budden",
-    text: null,
-    tweets: 1036
-  },
-  {
-    id: 6,
-    tag: "Trident",
-    text: null,
-    tweets: 6136
-  }
-];
 
 export default function() {
   return (
