+++ src/data/artefacts.js
@@ -1,0 +1,36 @@
+const publicURL = process.env.PUBLIC_URL;
+
+const artefacts = [
+  {
+    id: 1,
+    src: `${publicURL}/img/media-1.png`,
+    to: "/media_1"
+  },
+  {
+    id: 2,
+    src: `${publicURL}/img/media-2.png`,
+    to: "/media_2"
+  },
+  {
+    id: 3,
+    src: `${publicURL}/img/media-3.png`,
+    to: "/media_3"
+  },
+  {
+    id: 4,
+    src: `${publicURL}/img/media-4.png`,
+    to: "/media_4"
+  },
+  {
+    id: 5,
+    src: `${publicURL}/img/media-5.png`,
+    to: "/media_5"
+  },
+  {
+    id: 6,
+    src: `${publicURL}/img/media-6.png`,
+    to: "/media_6"
+  }
+];
+
+export default artefacts;
